% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Anova.R
\name{computeANOVA}
\alias{computeANOVA}
\title{Anova test for batch effect on individual gene expressions}
\usage{
computeANOVA(data, variable, is.log, n.cores)
}
\arguments{
\item{data}{S4 data object}

\item{variable}{character: The predictor variable to \code{lm} model. The variables included are 'Time' and 'Plate'}

\item{is.log}{logical: Checks if the S4 data has log values. It 'False', it converts data to log scale.}

\item{n.cores}{The number of cores to use, i.e. at most how many child processes will be run simultaneously. Must be
at least one, and parallel computing requires at least two cores.}
}
\value{
A S3 data frame. The output contains the Anova test (F test) scores corresponding to all genes in S4 data object.
}
\description{
This function is a part of the data analysis functionality of \code{tcgaCleaneR}. It use ANOVA F statistics to
summarize the effects of a qualitative source of unwanted variation (e.g. batches) on the expression levels of
individual genes. Unwanted Variations such as Plate effect and Time effect can be analysed by this test.
}
\examples{
\dontrun{
computeANOVA(data = brca.data, variable = "Time", is.log = FALSE, n.cores = 1)
computeANOVA(data = brca.data, variable = "Plate", is.log = FALSE, n.cores = 1)
}
}
