% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{computePCA}
\alias{computePCA}
\title{Generate PCA}
\usage{
computePCA(data, nPcs, is.log)
}
\arguments{
\item{data}{S4 data object}

\item{nPcs}{numeric: Number of PCs that needs to be generated}

\item{is.log}{logical: Checks if the S4 data has log values. If 'False', it converts data to log scale.}
}
\value{
A List of S4 class containing n PCs. For all three \code{assays()}, two values are returned. \code{sing.val} contains singular values with \code{u} containing PCs. \code{variation} contains the variation of each PC.
}
\description{
This function is a part of the data analysis functionality of \code{tcgaCleaneR}. It performs PCA using SVD
algorithm (\code{runSVD()}) on the \code{SummarizedExperiment} class TCGA Cancer Data across all \code{assays()} and
generate PCs.
}
\examples{
\dontrun{
computePCA(data = brca.data, nPcs = 10, is.log = FALSE)
}

}
