% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Anova.R
\name{f.test}
\alias{f.test}
\title{Anova test for year and plate effects}
\usage{
f.test(data, variable, is.log, n.cores)
}
\arguments{
\item{data}{S4 data object}

\item{variable}{character: The predictor variable to \code{lm} model. The variables included are 'Time' and 'Plate'}

\item{is.log}{logical: Checks if the S4 data has log values. It 'False', it converts data to log scale.}

\item{n.cores}{The number of cores to use, i.e. at most how many child processes will be run simultaneously. Must be at least one, and parallelization requires at least two cores.}
}
\value{
A S3 data frame. The output contains the Anova test (F test) scores corresponding to all genes in S4 data object.
}
\description{
This function is a part of the data analysis functionality of tgcapkg. It helps to perform the Anova test to analyse the variation effect due to time and plate on TCGA Cancer data.
}
\examples{
\dontrun{
f.test(data = brca.data, variable = "Plate", is.log = FALSE, n.cores = 2)
f.test(data = brca.data, variable = "Time", is.log = FALSE, n.cores = 2)
}
}
