% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_genes.R
\name{filter.genes}
\alias{filter.genes}
\title{Gene Filter}
\usage{
filter.genes(data,gene.type)
}
\arguments{
\item{data}{Input TGCA Dataset.}

\item{gene.type}{A character vector of items.}
}
\value{
S4 data object
}
\description{
This function is a part of the data wrangling functionality of tgcapkg.    It allows user to input the TGCA dataset and the required genes to filter the input data based on genes.
}
\examples{
filter.genes(data=brca.data,gene.type=c("protein_coding","lncRNA"))
\dontrun{
filter.genes(data=brca.data,gene.type=c("protein_coding"))
}
}
