% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_low_genes.R
\name{filter.low.genes}
\alias{filter.low.genes}
\title{Low Count Genes Filter}
\usage{
filter.low.genes(data, gene_count, sample_size)
}
\arguments{
\item{data}{Input TGCA Dataset.}

\item{gene_count}{numeric: gene count threshold}

\item{sample_size}{numeric: sample size threshold}
}
\value{
S4 data object
}
\description{
This function is a part of the data wrangling functionality of tgcapkg.    It allows user to input the TGCA dataset and the threshold for the minimum gene count and sample count.
}
\examples{
filter.low.genes(data=brca.se.data.temp,gene_count = 20,sample_size = 200)
}
