% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_low_genes.R
\name{filterLowExprGenes}
\alias{filterLowExprGenes}
\title{Low Count Genes Filter}
\usage{
filterLowExprGenes(data, gene_count, sample_size)
}
\arguments{
\item{data}{SummarizedExperiment S4 class Dataset. E.g. TCGA Dataset.}

\item{gene_count}{numeric: gene count threshold}

\item{sample_size}{numeric: sample size threshold}
}
\value{
S4 data object
}
\description{
This function is a part of the data wrangling functionality of \code{tcgaCleaneR}.
It allows user to input the \code{SummarizedExperiment} S4 class Cancer Dataset (e.g. TCGA dataset) and the threshold
for the minimum gene count and sample count.
}
\examples{

filterLowExprGenes(data=brca.data,gene_count = 20,sample_size = 200)

}
