% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library_size_filter.R
\name{filterSamplesByLibSize}
\alias{filterSamplesByLibSize}
\title{Filter Samples based on Library Size}
\usage{
filterSamplesByLibSize(data, ls_cutoff)
}
\arguments{
\item{data}{SummarizedExperiment S4 class Dataset. E.g. TCGA Dataset.}

\item{ls_cutoff}{numeric: library size threshold}
}
\value{
S4 data object
}
\description{
This function is a part of the data wrangling functionality of \code{tcgaCleaneR}.
It allows user to handle the bias in \code{SummarizedExperiment} S4 class Cancer Dataset (e.g. TCGA dataset) due to
library size by filtering out the samples with sample size greater than the threshold. Using \code{plotLibSize},
user can determine the threshold.
}
\examples{

filterSamplesByLibSize(data = brca.data, ls_cutoff = 17.5)

}
