% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purity_filter.R
\name{filterSamplesByPurity}
\alias{filterSamplesByPurity}
\title{Filter Samples based on Tumor Purity}
\usage{
filterSamplesByPurity(data, purity_cutoff)
}
\arguments{
\item{data}{SummarizedExperiment S4 class Dataset. E.g. TCGA Dataset.}

\item{purity_cutoff}{numeric: Sample purity cutoff}
}
\value{
S4 data object
}
\description{
This function is a part of the data wrangling functionality of \code{tcgaCleaneR}.
It allows user to handle sample purity in \code{SummarizedExperiment} S4 class Cancer Dataset (e.g. TCGA dataset) by
filtering out the samples that are above a specific purity threshold.
}
\examples{

filterSamplesByPurity(data= brca.data,purity_cutoff= 0.496)

}
