% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_genes.R
\name{gene.filter}
\alias{gene.filter}
\title{Gene Filter}
\usage{
gene.filter(data,gene.type)
}
\arguments{
\item{data}{SummarizedExperiment S4 class Dataset. E.g. TCGA Dataset.}

\item{gene.type}{A character vector of items.}
}
\value{
S4 data object
}
\description{
This function is a part of the data wrangling functionality of tgcapkg.
It allows user to input the \code{SummarizedExperiment} S4 class Cancer Dataset (e.g. TCGA dataset) and the required
genes to filter data based on genes.
}
\examples{

gene.filter(data=brca.data,gene.type=c("protein_coding","lncRNA"))
\dontrun{
gene.filter(data=brca.data,gene.type=c("protein_coding"))
}
}
