% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruv_prps.R
\name{get.prps}
\alias{get.prps}
\title{Generate RUV-III PRPS data}
\usage{
get.prps(
  data,
  batch,
  biology,
  purity,
  include.ls,
  include.purity,
  n.ls.ps,
  n.sample.ls,
  n.sample.batch,
  n.sample.purity
)
}
\arguments{
\item{data}{S4 data object}

\item{batch}{character: Batch effect factors. In current package version batch can be given values like 'Year', 'Plate' or both.}

\item{biology}{character: Biology of cancer type. In current package version biology for cancer type Breast, Rectum & Colon is considered.}

\item{purity}{character: Purity column name}

\item{include.ls}{logical: Do we need to consider library size in creating pseudo samples.}

\item{include.purity}{logical: Do we need to consider purity in creating pseudo samples.}

\item{n.ls.ps}{numeric: Number of samples for library size considered within plate.}

\item{n.sample.ls}{numeric: Minimum number of samples within library size pseudo sample size.}

\item{n.sample.batch}{numeric: Number of samples per batch.}

\item{n.sample.purity}{numeric: Number of samples for purity.}
}
\value{
A S4 list object with the Pseudo replicate for pseudo samples for different batches.
}
\description{
This function is a part of the data analysis functionality of \code{tgcapkg}. It creates pseudo samples for on Library size, Batches and Purity.
}
\examples{
\dontrun{
get.prps(data=brca.data, batch=c('Year', 'Plates'), biology='biology', purity='Purity_singscore', include.ls=TRUE, include.purity=TRUE, n.ls.ps=10, n.sample.batch=3, n.sample.purity=3, n.sample.ls=3)
get.prps(data=brca.data, batch=c('Year', 'Plates'), biology='biology', purity=NULL, include.ls=TRUE, include.purity=FALSE, n.ls.ps=10, n.sample.batch=3, n.sample.purity=0, n.sample.ls=3)
}
}
