% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library_size_filter.R
\name{library.size.fil}
\alias{library.size.fil}
\title{Filter Samples based on Library Size}
\usage{
library.size.fil(data, ls_cutoff)
}
\arguments{
\item{data}{Input TGCA Dataset.}

\item{ls_cutoff}{numeric: library size threshold}
}
\value{
S4 data object
}
\description{
This function is a part of the data wrangling functionality of tgcapkg.    It allows user to handle the bias in TCGA Cancer data due to library size by filtering out the samples with sample size greater than the threshold. Using \code{library.size}, user can determine the threshold.
}
\examples{
library.size.fil(data = brca.data, ls_cutoff = 22.5)
}
