% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_low_genes.R
\name{low.genes.filter}
\alias{low.genes.filter}
\title{Low Count Genes Filter}
\usage{
low.genes.filter(data, gene_count, sample_size)
}
\arguments{
\item{data}{Input TGCA Dataset.}

\item{gene_count}{numeric: gene count threshold}

\item{sample_size}{numeric: sample size threshold}
}
\value{
S4 data object
}
\description{
This function is a part of the data wrangling functionality of tgcapkg.    It allows user to input the TGCA dataset and the threshold for the minimum gene count and sample count.
}
\examples{

low.genes.filter(data=brca.data,gene_count = 20,sample_size = 200)

}
