% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc_reg_corr.R
\name{pca.corr}
\alias{pca.corr}
\title{Correlation Analysis between PCs and variation types}
\usage{
pca.corr(pca.data, data, type, nPCs)
}
\arguments{
\item{pca.data}{list: PCA output from \code{get.pca}.}

\item{data}{S4 data object}

\item{type}{character: The response variable to \code{lm} model. groups included are 'librarysize', 'purity' and 'time'.}

\item{nPCs}{numeric: Number of PCs that needs to be used for regression}
}
\value{
Linear Plot the compares the correlation between library size (or Purity, time) and PCs across three datasets. When output is stored in a object the user can also access values used to plot the linear graphs.
}
\description{
This function is a part of the data analysis functionality of tgcapkg. It helps user to run regression between bias in TCGA RNA-seq like librarysize and purity with PCs from \code{get.pca}. The output is a linear plot that compares the three \code{assays} in \code{SummarizedExperiment} TGCA Cancer data across n PCs and R-sq. It also runs vector correlation \code{stats::cancor} between time and n PCs with the same linear output explaining variation explained by different variation types.
}
\examples{
\dontrun{
pca.corr(pca.data, data = brca.data, type = "purity", nPCs = 10)
df <- pca.corr(pca.data, data = brca.data, type = "time", nPCs = 8)
df
}
}
