% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_PCA.R
\name{pca.plot}
\alias{pca.plot}
\title{PCA Visualization}
\usage{
pca.plot(pca.data, data, group, plot_type, npcs)
}
\arguments{
\item{pca.data}{list: PCA output from \code{get.pca}.}

\item{data}{S4 data object}

\item{group}{character: Color code PCs based on group. groups included are 'Time', 'Tissue', 'Plate', 'Batch', 'TSS', 'Center'}

\item{plot_type}{character: Plot type}

\item{npcs}{numeric: Number of PCs that needs to be plotted}
}
\value{
Density Plot, Box Plot
}
\description{
This function is a part of the data analysis functionality of tgcapkg. It helps to visualize the PCs from \code{get.pca}.
}
\examples{
\dontrun{
pca.plot(pca.data, data = brca.data, group = "Time", plot_type = "DensityPlot", npcs = 3)
pca.plot(pca.data = df6, data = brca.data, group = "Plate", plot_type = "BoxPlot", npcs = 3)
}
}
