% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library.size.R
\name{plotLibSize}
\alias{plotLibSize}
\title{Library Size}
\usage{
plotLibSize(data, plot_type)
}
\arguments{
\item{data}{SummarizedExperiment S4 class Dataset. E.g. TCGA Dataset.}

\item{plot_type}{character: Plot type}
}
\value{
Scatter Plot, Box plot
}
\description{
This function is a part of the data analysis functionality of \code{tcgaCleaneR}.
It allows user to analyze the Library Size bias, a technical bias in the TCGA RNA-seq.
The user can input the \code{SummarizedExperiment} S4 class Cancer Dataset (e.g. TCGA dataset) and the type of plot
to analyse variations in library size across years and samples.
}
\examples{
plotLibSize(data = brca.data, plot_type = "Scatterplot")
\dontrun{
plotLibSize(data = brca.data, plot_type = "Boxplot")
}
}
