% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_PCA.R
\name{plotPC}
\alias{plotPC}
\title{PCA Visualization}
\usage{
plotPC(pca.data, data, group, plot_type, pcs.no)
}
\arguments{
\item{pca.data}{list: PCA output from \code{computePCA}.}

\item{data}{S4 data object}

\item{group}{character: Color code PCs based on group. groups included are 'Time', 'Tissue', 'Plate', 'TSS', 'Center'}

\item{plot_type}{character: Plot type}

\item{pcs.no}{numeric vector: PCs that needs to be plotted}
}
\value{
Density Plot, Box Plot
}
\description{
This function is a part of the data analysis functionality of \code{tcgaCleaneR}. It helps to visualize the PCs from \code{get.pca}.
}
\examples{
\dontrun{
plotPC(pca.data, data = brca.data, group = "Time", plot_type = "DensityPlot", pcs.no = c(1,2,3))
plotPC(pca.data = df6, data = brca.data, group = "Plate", plot_type = "BoxPlot", pcs.no = c(1,2,4))
}
}
