% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruv_prps_plot.R
\name{plotPRPS}
\alias{plotPRPS}
\title{RUV-III PRPS Map}
\usage{
plotPRPS(data, n = 3)
}
\arguments{
\item{data}{S4 data object}

\item{n}{numeric: Minimum no. of samples needed to make pseudo sample. By default it consider 3 samples.}
}
\value{
PRPS Map which can be used to show how the biological variation are distributed across batches.
}
\description{
This function is a part of the data analysis functionality of \code{tcgaCleaneR}. It helps to map the pseudo samples
for different biology in a cancer type replicated across batches. The replication of of pseudo samples across batches
are known as Pseudo Replicates.
}
\examples{
\dontrun{
plotPRPS(data = brca.data)
}
}
