% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RUV_III.R
\name{runRUV_III_PRPS}
\alias{runRUV_III_PRPS}
\title{Generate RUV-III Data Object}
\usage{
runRUV_III_PRPS(
  ruv.data,
  ruv.rep,
  ncg.set,
  k = NULL,
  eta = NULL,
  include.intercept = TRUE,
  average = FALSE,
  fullalpha = NULL,
  return.info = FALSE,
  inputcheck = TRUE
)
}
\arguments{
\item{ruv.data}{S4 data object for RUV-III: A S4 data object with combined data including the row count from original filtered data using assay \code{HTseq_counts}, prps data for batch and library size. This data needs to be further converted to log scale and transposed.}

\item{ruv.rep}{S4 data matrix for RUV-III: A S4 data object that has been generated using \code{replicate.matrix} functionality from \code{ruv} package. This helps ruv to identify replicate samples.}

\item{ncg.set}{logical object: Set of Negative Controlled genes.}

\item{k}{Integer scalar specifying the number of unwanted factors to use. Default is NULL. Currently Value 1 represents the library size, 2 represents purity and 3 is time variation.}

\item{eta}{Gene-wise (as opposed to sample-wise) covariates. A matrix with n columns for \code{ruv::RUV1}. Default is NULL.}

\item{include.intercept}{Add an intercept term to eta if it does not include one already for \code{ruv::RUV1}. Default is True.}

\item{average}{Default is False.}

\item{fullalpha}{To perform RUV-III calculation. Default is NULL.}

\item{return.info}{logical: Do you want all the information related to RUV-III object. False gives all information whereas True gives only the}

\item{inputcheck}{logical: Check the inputs to identify if ruv.data contains missing values or infinite values.}
}
\value{
Based on the return.info we get either a S4 list will all information related to RUV-III object or just the RUV-III result.
}
\description{
This function is a part of the data analysis functionality of \code{tcgaCleaneR}. It captures both the uses PRPS values from library \code{tcgaCleaneR} combined with row counts and run SVD algorithm (\code{runSVD()}) from \code{BiocSingular} on the combined dataset. The function uses RUV-I algorithm from \code{ruv} as a pre-processing step to RUV-III.
}
\examples{
\dontrun{
runRUV_III_PRPS(ruv.data = ruv.data, ruv.rep = ruv.rep, ncg.set = ncg.set, k=1, return.info = TRUE)
}
}
