% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_genes.R
\name{gene.filter}
\alias{gene.filter}
\title{Gene Filter}
\usage{
gene.filter(data,gene.type)
}
\arguments{
\item{data}{Input TGCA Dataset.}

\item{gene.type}{A character vector of items.}
}
\value{
S4 data object
}
\description{
This function is a part of the data wrangling functionality of tgcapkg.
It allows user to input the TCGA dataset and the required genes to filter the input data based on genes.
}
\examples{

gene.filter(data=brca.data,gene.type=c("protein_coding","lncRNA"))
gene.filter(data=brca.data,gene.type=c("protein_coding"))

}
