% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library.size.R
\name{library.size}
\alias{library.size}
\title{Library Size}
\usage{
library.size(data, plot_type)
}
\arguments{
\item{data}{Input TGCA Dataset.}

\item{plot_type}{character: Plot type}
}
\value{
Scatter Plot, Box plot
}
\description{
This function is a part of the data analysis functionality of tgcapkg.    It allows user to analyze the Library Size bias, a technical bias in the TCGA RNA-seq.    The user can input the TGCA dataset and the type of plot to analyse variations in library size across years and samples.
}
\examples{
library.size(data = brca.data, plot_type = "Scatterplot")
\dontrun{
library.size(data = brca.data, plot_type = "Boxplot")
}
}
