% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_design.R
\name{study.design}
\alias{study.design}
\title{Study Design Plot}
\usage{
study.design(data)
}
\arguments{
\item{data}{S4 data object: Input TGCA Dataset}
}
\value{
Heat Map Plot
}
\description{
This function is a part of the data analysis functionality of tgcapkg. It helps to plot a Heat Map \code{ComplexHeatmap} for some sample variables and present a dashboard level analysis of TCGA data.
}
\examples{

study.design(data = brca.data)

}
